Started implementation of LLVM codegen. Also added a small runtime for Monga.

+++ src/libmonga/include/mon_literal.h
@@ -25,7 +25,7 @@ typedef struct {
 
         Mon_Char character;
         double real;
-        long integer;
+        long long integer;
     };
 
 } Mon_Literal;
